Changed DataStore to use blocks.
This is a bit of a hack.

The previous backup mechanism was causing quadratic slowdown, which
was obviously causing problems. We now write things in blocks, which
is fine. However, the new behavior makes the DataStore class a little
bit wonky to use and the internals are gross. I should fix this in
the future, but for now, at least we can run.