Removed assert statements from MELD kernel
There were numerous calls to assert that
energies were finite. Assert is not supported
on os x, so these calls have been removed.

This _may_ result in hangs when NaNs are
encountered. I believe this is due to
issues encountered when sorting the
restraints. If this turns out to be the case,
we will address it by other means besides
an assert.