Implemented explicit solvent calculations
When creating system options, you can now
specify 'implicit' or 'explicit'.

You can now load a system directly from
prmtop and mdcrd files.

Periodic boundary conditions are supported
using either cutoffs or PME.

Pressure coupling is supported, but the pressure
is not used during exchanges, which should be
a very minor effect.

It is likely that many of the different types of
MELD restraints do not support PBC, so they
likely will not work, especially when adding
restraints between molecules.

Replica exchange is likely to perform extremely
poorly when varying the temperature. In the future,
we will implement REST2 or similar.