Change type of energyBuffer to mixed
Previously, energyBuffer was defined as type
real, which is incorrect. In single precision,
real is defined as type float, while in mixed,
it is defined as double. However, the actual
energy buffers are floats for both single and
mixed. The switch to mixed precision in
a727279b4449454e34fb4cf51bc07686b44cb0f5 led
the MELD energies to be wildly incorrect.