Replace parallel cumulative sum with serial version.
Replaced a parallel algorithm with a serial one when
evaluating collections. The previous algorithm was
introduced in e71daa01. It gives correct results on a
GTX 980, but not on a Tesla 2070. This is a temporary
fix that might have minor performance implications and
should probably be swapped for a version of the
parallel algorithm that is correct on all hardware.